requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.46
pypdf
faiss-cpu
sentence-transformers
huggingface-hub
